fix(test): clean up skill-aware llm runner
