+++ backend/src/agent-skills/structure-type/README.md
@@ -22,4 +22,6 @@ Testing conventions:
 - These colocated tests must stay deterministic. They can use `llm = null` or stubbed responses, but they must not call a real external LLM provider.
 - Real provider-backed skill cases live next to the skill under `backend/src/agent-skills/**/__llm_tests__/`.
 - The repo-root `tests/llm-integration/` folder only keeps shared runner/helpers plus fixture discovery.
-- Every real LLM fixture must declare `skillId`, so tests can be filtered per skill with `node tests/runner.mjs llm-integration --skill <skillId>`.
+- New real LLM fixtures should use the v2 `family + scenarios + variants` shape so each scenario can declare explicit `enabledSkillIds`, `fallbackPolicy`, and `expect`.
+- Legacy v1 fixtures with top-level `skillId` and `testCases[]` are still supported, but new coverage should use v2.
+- Targeted runs can filter by family, variant, and scenario, for example `node tests/runner.mjs llm-integration --family frame --variant specific`.
