+++ tests/llm-integration/README.md
@@ -9,12 +9,19 @@
 ## Fixture Rules
 
 - Real LLM cases live next to the owning skill under `backend/src/agent-skills/**/__llm_tests__/`.
-- Every case in those colocated fixture files must declare `skillId`.
-- `skillId` is used for targeted runs like:
+- Skill-aware fixtures should use the v2 shape:
+  - document-level `family`
+  - `scenarios[]`
+  - `scenarioId`
+  - per-variant `enabledSkillIds`, `fallbackPolicy`, and `expect`
+- Legacy v1 fixtures with `testCases[]` and `assertions` are still supported, but new work should use v2.
+- Targeted runs can filter by `family`, `variant`, and `scenario`. `--skill` remains a compatibility alias for `--family`.
 
 ```bash
-node tests/runner.mjs llm-integration --skill frame
-node tests/runner.mjs llm-integration extraction --skill beam
+node tests/runner.mjs llm-integration --family frame
+node tests/runner.mjs llm-integration pipeline --family frame --variant specific
+node tests/runner.mjs llm-integration extraction --family beam --scenario beam-basic
+node tests/runner.mjs llm-integration pipeline --family frame --variant generic --output tests/.artifacts/frame-generic.json
 ```
 
 ## Recommended Commands
@@ -28,4 +35,5 @@ npm run test:llm
 npm run test:llm:beam
 npm run test:llm:frame
 npm run test:llm:portal-frame
+node --test tests/llm-integration/lib/*.test.cjs
 ```
